+++ src/main/java/net/frozenblock/wilderwild/tag/WWBiomeTags.java
@@ -147,6 +147,7 @@ public final class WWBiomeTags {
 	public static final TagKey<Biome> HAS_CATTAIL = bind("feature/has_cattail");
 	public static final TagKey<Biome> HAS_CATTAIL_UNCOMMON = bind("feature/has_cattail_uncommon");
 	public static final TagKey<Biome> HAS_CATTAIL_COMMON = bind("feature/has_cattail_common");
+	public static final TagKey<Biome> HAS_CATTAIL_MUD = bind("feature/has_cattail_mud");
 	public static final TagKey<Biome> HAS_SEEDING_DANDELION = bind("feature/has_seeding_dandelion");
 	public static final TagKey<Biome> HAS_COMMON_SEEDING_DANDELION = bind("feature/has_common_seeding_dandelion");
 	public static final TagKey<Biome> HAS_RARE_SEEDING_DANDELION = bind("feature/has_rare_seeding_dandelion");